+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/recycle-bin/tree/data/index.ts
@@ -1,0 +1,2 @@
+export { UmbDocumentRecycleBinTreeRepository } from './document-recycle-bin-tree.repository.js';
+export { type UmbDocumentRecycleBinTreeStore } from './document-recycle-bin-tree.store.js';
