+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/recycle-bin/tree/index.ts
@@ -1,2 +1,1 @@
-export { UmbDocumentRecycleBinTreeRepository } from './document-recycle-bin-tree.repository.js';
-export { type UmbDocumentRecycleBinTreeStore } from './document-recycle-bin-tree.store.js';
+export * from './data/index.js';
