+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/recycle-bin/tree/manifests.ts
@@ -1,24 +1,9 @@
 import { UMB_DOCUMENT_RECYCLE_BIN_ROOT_ENTITY_TYPE } from '../constants.js';
-import {
-	UMB_DOCUMENT_RECYCLE_BIN_TREE_ALIAS,
-	UMB_DOCUMENT_RECYCLE_BIN_TREE_REPOSITORY_ALIAS,
-	UMB_DOCUMENT_RECYCLE_BIN_TREE_STORE_ALIAS,
-} from './constants.js';
+import { UMB_DOCUMENT_RECYCLE_BIN_TREE_ALIAS, UMB_DOCUMENT_RECYCLE_BIN_TREE_REPOSITORY_ALIAS } from './constants.js';
 import { manifests as reloadTreeItemChildrenManifests } from './reload-tree-item-children/manifests.js';
+import { manifests as dataManifests } from './data/manifests.js';
 
 export const manifests: Array<UmbExtensionManifest> = [
-	{
-		type: 'repository',
-		alias: UMB_DOCUMENT_RECYCLE_BIN_TREE_REPOSITORY_ALIAS,
-		name: 'Document Recycle Bin Tree Repository',
-		api: () => import('./document-recycle-bin-tree.repository.js'),
-	},
-	{
-		type: 'treeStore',
-		alias: UMB_DOCUMENT_RECYCLE_BIN_TREE_STORE_ALIAS,
-		name: 'Document Recycle Bin Tree Store',
-		api: () => import('./document-recycle-bin-tree.store.js'),
-	},
 	{
 		type: 'tree',
 		kind: 'default',
@@ -46,5 +31,6 @@ export const manifests: Array<UmbExtensionManifest> = [
 			headline: '#general_recycleBin',
 		},
 	},
+	...dataManifests,
 	...reloadTreeItemChildrenManifests,
 ];
